requirements.txt
pandas>=1.3.0
numpy>=1.20.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
transformers>=4.20.0
sentence-transformers>=2.2.0
torch>=1.10.0
pyyaml>=6.0
pathlib>=1.0.1
